1. Changed directory structure (introduced /lib for all library files) 2. Changed pattern for request and signal taking go as reference implementation 3. Request and Response added for handler and http calls
