+++ lib/handler_req.rb
@@ -1,0 +1,5 @@
+module Gilmour
+  class HandlerReq
+    attr_accessor :data
+  end
+end
