+++ lib/handler_resp.rb
@@ -1,0 +1,5 @@
+module Gilmour
+  class HandlerResp
+    attr_accessor :data
+  end
+end
