+++ lib/response.rb
@@ -1,0 +1,10 @@
+module Gilmour
+  class Response
+    attr_reader :code, :data
+
+    def initialize(body, code)
+      @data = body
+      @code = code
+    end
+  end
+end
