--- request.rb
@@ -1,3 +1,0 @@
-class Request
-  attr_accessor :data
-end
