--- responder.rb
@@ -1,6 +1,0 @@
-class Response
-  attr_reader :data
-  def set_data(data)
-    @data = data
-  end
-end
